[BEAM-8376] Google Cloud Firestore Connector - Add Firestore v1 Write Operations
Entry point for accessing Firestore V1 read methods is `FirestoreIO.v1().write()`.

Currently supported write RPC methods:
* `BatchWrite`

### Unit Tests

No external dependencies are needed for this suite

A large suite of unit tests have been added to cover most branches and error
scenarios in the various components. Test for input validation and bounds
checking are also included in this suite.

### Integration Tests

Integration tests for each type of RPC is present in
`org.apache.beam.sdk.io.gcp.firestore.it.FirestoreV1IT`. All of these tests
leverage `TestPipeline` and verify the expected Documents/Collections are all
operated on during the test.

+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/firestore/FirestoreProtoHelpers.java
@@ -25,11 +25,9 @@ final class FirestoreProtoHelpers {
     return newWrite(1);
   }
 
-  static Write newWrite(int i) {
+  static Write newWrite(long i) {
     Write.Builder writeBuilder = Write.newBuilder();
-    writeBuilder
-        .getUpdateBuilder()
-        .setName(String.format("doc-%03d", i));
+    writeBuilder.getUpdateBuilder().setName(String.format("doc-%012d", i));
     return writeBuilder.build();
   }
 }
